+++ boot.php
@@ -153,8 +153,8 @@ if (rex::isFrontend() || rex::isBackend()) {
     if (!class_exists('consent_manager_inline')) {
         require_once __DIR__ . '/lib/consent_manager_inline.php';
     }
-    if (!class_exists('rex_api_consent_manager_inline_log')) {
-        require_once __DIR__ . '/lib/api_inline_log.php';
+    if (!class_exists(FriendsOfRedaxo\ConsentManager\ApiInlineLog::class)) {
+        require_once __DIR__ . '/lib/ApiInlineLog.php';
     }
 }